dopracowanie MediaView, ikonki statusów

+++ AmeCommon/MediaTasks/BaseMediaTask.cs
@@ -1,6 +1,7 @@
 ﻿using System.IO;
 using AmeCommon.MediaTasks.TaskHandlers;
 using Microsoft.VisualBasic.FileIO;
+using System;
 
 namespace AmeCommon.MediaTasks
 {
@@ -8,15 +9,21 @@ namespace AmeCommon.MediaTasks
     {
         public abstract void Execute();
 
-        protected void MoveAllDirectoryContent(string absoluteSource, string absoluteTarget)
+        protected void MoveAllDirectoryContent(string absoluteSource, string absoluteTarget, bool deleteSourcDir)
         {
+            if (!Directory.Exists(absoluteSource))
+                return;
             MoveAllFiles(absoluteSource, absoluteTarget);
             MoveAllChildDirectories(absoluteSource, absoluteTarget);
-            Directory.Delete(absoluteSource);
+            if (deleteSourcDir)
+                Directory.Delete(absoluteSource);
         }
 
         protected void MoveAllChildDirectories(string absoluteSource, string absoluteTarget)
         {
+            if (!Directory.Exists(absoluteSource))
+                return;
+
             foreach (var childSourceDir in Directory.GetDirectories(absoluteSource))
             {
                 var destDir = Path.Combine(absoluteTarget, Path.GetFileName(childSourceDir) ?? "");
@@ -26,6 +33,9 @@ namespace AmeCommon.MediaTasks
 
         protected void MoveAllFiles(string absoluteSource, string absoluteTarget, string searchPattern=null)
         {
+            if (!Directory.Exists(absoluteSource))
+                return;
+
             Directory.CreateDirectory(absoluteTarget);
 
             var sourceFiles = searchPattern != null
@@ -36,8 +46,58 @@ namespace AmeCommon.MediaTasks
             foreach (var sourceFile in sourceFiles)
             {
                 var destFile = Path.Combine(absoluteTarget, Path.GetFileName(sourceFile) ?? "");
-                File.Move(sourceFile, destFile);
+                MoveFile(sourceFile, destFile);
             }
+        }
+
+        private bool AreFilesSameByContent(string sourceFile, string destFile)
+        {
+            FileInfo src = new FileInfo(sourceFile);
+            FileInfo dst = new FileInfo(destFile);
+            if (src.Length != dst.Length)
+                return false;
+
+            using (FileStream fs1 = src.OpenRead())
+            using (FileStream fs2 = dst.OpenRead())
+            {
+                while (true)
+                {
+                    var b1 = fs1.ReadByte();
+                    var b2 = fs2.ReadByte();
+                    if (b1 != b2)
+                        return false;
+                    if (b1 < 0)
+                        return true;
+                }
+            }
+        }
+
+        private void MoveFile(string sourceFile, string destFile)
+        {
+            if (File.Exists(destFile))
+            {
+                if (AreFilesSameByContent(sourceFile, destFile))
+                    File.Delete(sourceFile);
+                else
+                    File.Move(sourceFile, GetAlternativeName(destFile));
+            }
+            else
+                File.Move(sourceFile, destFile);
+        }
+
+        private string GetAlternativeName(string destFile)
+        {
+            var path = Path.GetDirectoryName(destFile);
+            var ext = Path.GetExtension(destFile);
+            var name = Path.GetFileNameWithoutExtension(destFile);
+
+            for(int i=1; i<int.MaxValue; i++)
+            {
+                var newName = Path.Combine(path, name + "_" + i + ext);
+                if (!File.Exists(newName))
+                    return newName;
+            }
+            throw new ApplicationException();
         }
     }
 }
